+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 ---------
 
+0.2.0 (September 1, 2015)
+^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Changed ``qstring.nest`` to maintain the order of the given parameters in the
+  returned nested object by returning ``OrderedDict`` instead of ``dict``.
+
+
 0.1.0 (July 14, 2015)
 ^^^^^^^^^^^^^^^^^^^^^